Implement pawn pushes/captures, and begin implementation of knight moves
The current approach uses bitboards for the board representation, most
of which is contained in `crates/chess_engine/src/board.rs`. Some of the
board representation is still in `crates/chess_engine/src/engine.rs`, but
most of this should eventually be moved to `board.rs`.

The board representation can currently handle generating pawn moves (with
the exception of En Passant), and is in the early stages of generating
knight moves. Importantly, there is no concept of check as of yet. This
means the engine will happily play a move that leaves it in check, or
ignores that it is in check already. This will obviously need to be
implemented at some point.